전체 목록 + 클릭 시 상세 정보 + 예매 누르면 예매 화면 (데이터 연동 아직 ㄴㄴ) 수정 예정

+++ src/main/java/mybatis/vo/Movie.java
@@ -13,7 +13,7 @@ import lombok.ToString;
 @Builder
 public class Movie {
 	
-	private int mcode;				// 영화 코드	
+	private long mcode;				// 영화 코드	
 	private String mname;			// 영화 제목
 	private String genre;			// 영화 장르
 	private String director;			// 영화 감독
